refactor: 각 컴포넌트 내에서 props interface 정의하는 방식으로 리팩토링
