+++ src/Page/Client/ClientSelectList/PaymentModal/PrintReceipt.tsx
@@ -5,7 +5,8 @@ interface ReceiptProps {
   remain: number;
   handleReceipt: (receipt: boolean) => void;
 }
-const PrintReceipt: React.FC<ReceiptProps> = ({ remain, handleReceipt }) => {
+
+const PrintReceipt = ({ remain, handleReceipt }: ReceiptProps) => {
   return (
     <PaymentBox>
       <h2>결제가 완료되었습니다</h2>
